Remove genetic algorithm and any references to it.
Not sure why I ever did this instead of just using letter frequencies.
This seems to work better.

Thanks to Sjoerd Langkemper in a208ca2f58097b43cf3c593f9b3f8abc3fdb06a7
for giving the impotus to make this change.
