+++ app/src/test/resources/letters_fr_fr.txt
@@ -6,7 +6,7 @@
 #  - If twice, the 1st probability is incremented by 2, and the 2nd probability by 1, etc.
 #
 # When generating 100 4x4 boards, the following number of words were observed:
-#   Min: 12, mean: 77, max: 207, stddev: 38, score: 2944
+#   Min: 20, mean: 71, max: 192, stddev: 39
 #
 
 a 84290 20808 1496 23
